requirements.txt
discord.py
requests
os
sys
json